% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{$.influxdb}
\alias{$.influxdb}
\title{Call InfluxDB HTTP API}
\usage{
\method{$}{influxdb}(con, func)
}
\arguments{
\item{con}{an InfluxDB connection}

\item{func}{function to call}
}
\value{
parsed csv
}
\description{
Call InfluxDB HTTP API
}
