% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{convert_kv}
\alias{convert_kv}
\title{Convert a named list to key-value pairs}
\usage{
convert_kv(x, do_quote, do_escape, force_integer)
}
\arguments{
\item{x}{a named list}

\item{do_quote}{whether to quote values}

\item{do_escape}{whether to check and escape special characters}

\item{force_integer}{wheter to use integer data type}
}
\value{
a list of key-value paris
}
\description{
Convert a named list to key-value pairs
}
