% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{escape}
\alias{escape}
\title{Check and escape special characters}
\usage{
escape(x, esc_equal)
}
\arguments{
\item{x}{a vector of characters}

\item{esc_equal}{whether to escape equal signs}
}
\value{
a vector of characters
}
\description{
Check and escape special characters
}
