% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb}
\alias{influxdb}
\title{Create an InfluxDB connection}
\usage{
influxdb(scheme = c("http", "https"), host, port, user = NULL,
  pass = NULL, path = "/")
}
\arguments{
\item{scheme}{scheme to use}

\item{host}{hostname}

\item{port}{port number}

\item{user}{username}

\item{pass}{password}

\item{path}{API path}
}
\value{
an InfluxDB connection
}
\description{
Create an InfluxDB connection
}
