% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb_chkr}
\alias{influxdb_chkr}
\title{Check InfluxDB response. Parse error message and re-throw if any.}
\usage{
influxdb_chkr(r)
}
\arguments{
\item{r}{an httr response}
}
\value{
r itself
}
\description{
Check InfluxDB response. Parse error message and re-throw if any.
}
