% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb_credential}
\alias{influxdb_credential}
\title{Get credential from an InfluxDB connection object.}
\usage{
influxdb_credential(con)
}
\arguments{
\item{con}{an InfluxDB connection object}
}
\value{
a list of credential
}
\description{
Get credential from an InfluxDB connection object.
}
