% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb_csv}
\alias{influxdb_csv}
\title{Parse InfluxDB CSV query. Use data.table fread if possible.}
\usage{
influxdb_csv(r)
}
\arguments{
\item{r}{an httr response}
}
\value{
data.frame like object
}
\description{
Parse InfluxDB CSV query. Use data.table fread if possible.
}
