% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb_get}
\alias{influxdb_get}
\title{GET InfluxDB endpoint}
\usage{
influxdb_get(con, endpoint, query = NULL, httr_config = list(),
  parser = c("json", "csv"))
}
\arguments{
\item{con}{InfluxDB connection object}

\item{endpoint}{endpoint to POST}

\item{query}{query data}

\item{httr_config}{curl config passed to httr::POST}

\item{parser}{request json/csv}
}
\value{
httr response
}
\description{
GET InfluxDB endpoint
}
