% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb_ping}
\alias{influxdb_ping}
\title{InfluxDB ping}
\usage{
influxdb_ping(con)
}
\arguments{
\item{con}{an InfluxDB connection object}
}
\value{
list containing InfluxDB build and version
}
\description{
InfluxDB ping
}
