% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb_post}
\alias{influxdb_post}
\title{POST to InfluxDB endpoint}
\usage{
influxdb_post(con, endpoint, query = NULL, body = NULL,
  httr_config = list(), parser = c("json", "csv"))
}
\arguments{
\item{con}{InfluxDB connection object}

\item{endpoint}{endpoint to POST}

\item{query}{query data}

\item{body}{body data}

\item{httr_config}{curl config passed to httr::POST}

\item{parser}{request json/csv}
}
\value{
httr response
}
\description{
POST to InfluxDB endpoint
}
