% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{influxdb_query}
\alias{influxdb_query}
\title{Make InfluxDB query}
\usage{
influxdb_query(q, con, db, epoch, httr_config = list())
}
\arguments{
\item{q}{query}

\item{con}{an InfluxDB connection object}

\item{db}{database}

\item{epoch}{unit of epoch}

\item{httr_config}{additional httr curl config passed to httr::POST()}
}
\value{
a data.table
}
\description{
Make InfluxDB query
}
