% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{influxdb_write}
\alias{influxdb_write}
\alias{influxdb_write.default}
\alias{influxdb_write.character}
\title{Write data to InfluxDB}
\usage{
influxdb_write(data, con, db, precision, rp, consistency,
  httr_config = list(), ...)

\method{influxdb_write}{default}(data, con, db, precision, rp, consistency,
  httr_config = list(), ...)

\method{influxdb_write}{character}(data, con, db, precision, rp,
  consistency, httr_config = list(), ...)
}
\arguments{
\item{data}{data to write, a vector line protocals or a named list-like object}

\item{con}{a InfluxDB connection}

\item{db}{database to write to}

\item{precision}{precision of the epoch}

\item{rp}{retention policy name}

\item{consistency}{consistency}

\item{httr_config}{additional httr curl config passed to httr::POST()}

\item{...}{futher arguments passed to line_protocol()}
}
\value{
httr reponse
}
\description{
Write data to InfluxDB
}
