% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{kv_fmtstr}
\alias{kv_fmtstr}
\title{Generate format string for key=value pair from a reference named list.}
\usage{
kv_fmtstr(ref, key, esc, use_int = FALSE, fp_prec = NULL,
  quote_str = TRUE, as_str = FALSE, op = "=", sep = ",",
  tail = " ")
}
\arguments{
\item{ref}{a named list, or a matrix-like object}

\item{key}{key}

\item{esc}{a function to escape special characters in key}

\item{use_int}{whether to use integer data type}

\item{fp_prec}{fp precision}

\item{quote_str}{whether to quote string value}

\item{as_str}{whether to print all values as string without quote}

\item{op}{operator}

\item{sep}{seperator}

\item{tail}{tailing string}
}
\value{
a character vector
}
\description{
Generate format string for key=value pair from a reference named list.
}
