% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{line_protocol}
\alias{line_protocol}
\title{Convert data to InfluxDB line protocol}
\usage{
line_protocol(data = list(), measurement, tag = NULL, field = NULL,
  time = NULL, do_escape = FALSE, force_integer = FALSE,
  epoch = c("ns", "u", "ms", "s", "m", "h"))
}
\arguments{
\item{data}{a named list-like object}

\item{measurement}{name of measurement or column name of measurement in data}

\item{tag}{a named list of tag k-v, or column name of tags in data}

\item{field}{a named list of field k-v, or column name of fields in data}

\item{time}{a vector of POSIXct, or column name of time in data}

\item{do_escape}{whether to check special characters in key/value that needs to escape}

\item{force_integer}{whether to use integer data type}

\item{epoch}{unit of epoch a.k.a precision}
}
\value{
a vector of characters
}
\description{
Convert data to InfluxDB line protocol
}
\details{
This function checks whether measurement/tag/field/time are named.
 If not, these arguments are used as references to columns in data. Checking
 and escaping special characters are turned off by default to improve
 performance. If you have special characters in your data, set do_escape = TRUE
}
