% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.influxdb}
\alias{print.influxdb}
\title{Print an InfluxDB connection}
\usage{
\method{print}{influxdb}(x, ...)
}
\arguments{
\item{x}{an InfluxDB connection}

\item{...}{not used}
}
\value{
x, invisibly
}
\description{
Print an InfluxDB connection
}
