% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{quote}
\alias{quote}
\title{Check and quote values if requied}
\usage{
quote(x, do_escape, force_integer)
}
\arguments{
\item{x}{a vector of values}

\item{do_escape}{whether to check and escape special characters}

\item{force_integer}{whether to use integer data type}
}
\value{
a vector of values
}
\description{
Check and quote values if requied
}
