% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{influxdb}
\alias{influxdb}
\title{Create an InfluxDB connection}
\usage{
influxdb(host, port = 8086, user = NULL, pass = NULL,
  scheme = c("http", "https"), path = "/")
}
\arguments{
\item{host}{hostname}

\item{port}{port number}

\item{user}{username}

\item{pass}{password}

\item{scheme}{scheme to use}

\item{path}{API path}
}
\value{
an InfluxDB connection
}
\description{
Create an InfluxDB connection
}
