% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{line_protocol}
\alias{line_protocol}
\alias{line_protocol_gen}
\title{Convert data to InfluxDB line protocol}
\usage{
line_protocol(data, measurement, tag = NULL, field = NULL,
  time = NULL, chk_escape = FALSE, chk_na = FALSE, use_int = FALSE,
  fp_prec = NULL, epoch = c("ns", "u", "ms", "s", "m", "h"))

line_protocol_gen(ref, measurement, tag = NULL, field = NULL,
  time = NULL, chk_escape = FALSE, chk_na = FALSE, use_int = FALSE,
  fp_prec = NULL, epoch = c("ns", "u", "ms", "s", "m", "h"))
}
\arguments{
\item{data}{data to convert.}

\item{measurement}{measurement name, or column name of measurement in ref}

\item{tag}{column names of tags in ref}

\item{field}{column names of fields in ref, if NULL, all columns will be used as fields}

\item{time}{column name of time in ref}

\item{chk_escape}{whether to check special characters to escape}

\item{chk_na}{whether to check NA values}

\item{use_int}{whether to use integer data type}

\item{fp_prec}{precision of floating point data, if NULL, system default is used}

\item{epoch}{unit of epoch}

\item{ref}{a reference data structure}
}
\value{
a function that accepts same data structure as ref, or a character vector of line protocol
}
\description{
Convert data to InfluxDB line protocol
}
\details{
line_protocol() converts a named data structure to a character vector
of InfluxDB line protocol. It is a wrapper of line_protocol_gen(), which parses
reference data and returns a converter function that accepts same data structure.

 Remarks on NA values. if chk_na is TURE, line_protocol_gen() checks for NA
 values in fields ONLY and uses regex ",(?:(?!,|=).)*?=NA" to match/remove NA
 fields. However this comes with two caveat: "=NA" inside string field values
 are replace by random token @_=_N_A@ + a random number and later replaced back
 to =NA, if you have such values and accidentally with same random token,
 Secondly, as you may notice, the mentioned regex can't distinguish field
 keys with special characters such as special\,key\=NA=100 will cause problem.
}
\examples{
ref <- iris
f <- line_protocol_gen(ref, "iris", tag = "Species")
lp <- f(ref)
lp2 <- line_protocol(ref, "iris", tag = "Species")
stopifnot(all(lp == lp2))
}
